add export feature ++ make delete item easier to read
