+++ starsky/starsky.feature.export/Services/ExportService.cs
@@ -1,0 +1,195 @@
+using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using starsky.feature.export.Interfaces;
+using starsky.foundation.database.Helpers;
+using starsky.foundation.database.Interfaces;
+using starsky.foundation.database.Models;
+using starsky.foundation.platform.Helpers;
+using starsky.foundation.platform.Models;
+using starsky.foundation.storage.Interfaces;
+using starsky.foundation.storage.Storage;
+using starsky.foundation.thumbnailgeneration.Services;
+
+namespace starsky.feature.export.Services
+{
+	public class ExportService: IExport
+	{
+		
+		private readonly IQuery _query;
+		private readonly AppSettings _appSettings;
+		private readonly IStorage _iStorage;
+		private readonly IStorage _thumbnailStorage;
+		private readonly IStorage _hostFileSystemStorage;
+		private readonly StatusCodesHelper _statusCodeHelper;
+
+		public ExportService(	IQuery query, AppSettings appSettings, 
+			ISelectorStorage selectorStorage)
+		{
+			_appSettings = appSettings;
+			_query = query;
+			_iStorage = selectorStorage.Get(SelectorStorage.StorageServices.SubPath);
+			_thumbnailStorage = selectorStorage.Get(SelectorStorage.StorageServices.Thumbnail);
+			_hostFileSystemStorage = selectorStorage.Get(SelectorStorage.StorageServices.HostFilesystem);
+			_statusCodeHelper = new StatusCodesHelper();
+		}
+
+		// todo: !!!!!! RENAME!!!!
+		public Tuple<string, List<FileIndexItem>> CreateZip(string[] inputFilePaths, 
+			bool collections = true, 
+			bool thumbnail = false )
+		{
+			
+			// the result list
+			var fileIndexResultsList = new List<FileIndexItem>();
+
+			foreach ( var subPath in inputFilePaths )
+			{
+				var detailView = _query.SingleItem(subPath, null, collections, false);
+
+				if ( detailView?.FileIndexItem == null )
+				{
+					_statusCodeHelper.ReturnExifStatusError(new FileIndexItem(subPath), 
+						FileIndexItem.ExifStatus.NotFoundNotInIndex,
+						fileIndexResultsList);
+					continue;
+				}
+				
+				if ( !_iStorage.ExistFile(detailView.FileIndexItem.FilePath) )
+				{
+					_statusCodeHelper.ReturnExifStatusError(detailView.FileIndexItem, 
+						FileIndexItem.ExifStatus.NotFoundSourceMissing,
+						fileIndexResultsList);
+					continue; 
+				}
+				
+				// all filetypes that are exist > should be added 
+				
+				// todo: add filesystem check
+				
+				
+				
+				// var statusResults =
+				// 	new StatusCodesHelper(_appSettings).FileCollectionsCheck(detailView);
+				//
+				// // ignore readonly status
+				// if ( statusResults == FileIndexItem.ExifStatus.ReadOnly )
+				// 	statusResults = FileIndexItem.ExifStatus.Ok;
+				//
+				//
+				// var statusModel = new FileIndexItem();
+				// statusModel.SetFilePath(subPath);
+				// statusModel.IsDirectory = false;
+				//
+				// if(new StatusCodesHelper().ReturnExifStatusError(statusModel, statusResults, fileIndexResultsList)) continue;
+				//
+				//
+				
+
+				// if ( detailView == null ) throw new InvalidDataException("DetailView is null ~ " + nameof(detailView));
+
+				// Now Add Collection based images
+				var collectionSubPathList = detailView.GetCollectionSubPathList(detailView, collections, subPath);
+				foreach ( var item in collectionSubPathList )
+				{
+					var itemDetailView = _query.SingleItem(item, null, false, false).FileIndexItem;
+					itemDetailView.Status = FileIndexItem.ExifStatus.Ok;
+					fileIndexResultsList.Add(itemDetailView);
+				}
+			}
+
+			var isThumbnail = thumbnail ? "TN" : "SR"; // has:notHas
+			var zipHash = isThumbnail + GetName(fileIndexResultsList);
+			
+			return new Tuple<string, List<FileIndexItem>>(zipHash, fileIndexResultsList);
+		}
+		
+		/// <summary>
+		/// This list will be included in the zip
+		/// </summary>
+		/// <param name="fileIndexResultsList">the items</param>
+		/// <param name="thumbnail">add the thumbnail or the source image</param>
+		/// <returns>list of file paths</returns>
+		public List<string> CreateListToExport(List<FileIndexItem> fileIndexResultsList, bool thumbnail)
+		{
+			var filePaths = new List<string>();
+
+			foreach ( var item in fileIndexResultsList.Where(p => p.Status == FileIndexItem.ExifStatus.Ok).ToList() )
+			{
+				var sourceFile = _appSettings.DatabasePathToFilePath(item.FilePath);
+				var sourceThumb = Path.Join(_appSettings.ThumbnailTempFolder,
+					item.FileHash + ".jpg");
+
+				if ( thumbnail )
+					new Thumbnail(_iStorage, _thumbnailStorage).CreateThumb(item.FilePath, item.FileHash);
+
+				filePaths.Add(thumbnail ? sourceThumb : sourceFile); // has:notHas
+				
+				
+				// when there is .xmp sidecar file
+				if ( !thumbnail && ExtensionRolesHelper.IsExtensionForceXmp(item.FilePath) 
+				                && _iStorage.ExistFile(ExtensionRolesHelper.ReplaceExtensionWithXmp(item.FilePath)))
+				{
+					filePaths.Add(
+						_appSettings.DatabasePathToFilePath(ExtensionRolesHelper.ReplaceExtensionWithXmp(item.FilePath))
+					);
+				}
+				
+			}
+
+			return filePaths;
+		}
+		
+		/// <summary>
+		/// Get the filename (in case of thumbnail the source image name)
+		/// </summary>
+		/// <param name="filePaths">the full file paths </param>
+		/// <param name="thumbnail">copy the thumbnail (true) or the source image (false)</param>
+		/// <returns></returns>
+		public List<string> FilePathToFileName(List<string> filePaths, bool thumbnail)
+		{
+			var fileNames = new List<string>();
+			foreach ( var filePath in filePaths )
+			{
+				if ( thumbnail )
+				{
+					// We use base32 fileHashes but export 
+					// the file with the original name
+					
+					var thumbFilename = Path.GetFileNameWithoutExtension(filePath);
+					var subPath = _query.GetSubPathByHash(thumbFilename);
+					var filename = subPath.Split("/").LastOrDefault();
+					fileNames.Add(filename);
+					continue;
+				}
+				fileNames.Add(Path.GetFileName(filePath));
+			}
+			return fileNames;
+		}
+
+
+
+		/// <summary>
+		/// to create a unique name of the zip using c# get hashcode
+		/// </summary>
+		/// <param name="fileIndexResultsList">list of objects with filehashes</param>
+		/// <returns>unique 'get hashcode' string</returns>
+		public string GetName(List<FileIndexItem> fileIndexResultsList)
+		{
+			var tempFileNameStringBuilder = new StringBuilder();
+			foreach ( var item in fileIndexResultsList )
+			{
+				tempFileNameStringBuilder.Append(item.FileHash);
+			}
+			// to be sure that the max string limit
+			var shortName = tempFileNameStringBuilder.ToString().GetHashCode()
+				.ToString(CultureInfo.InvariantCulture).ToLower().Replace("-","A");
+			
+			return shortName;
+		}
+	}
+}
