+++ starsky/starsky.sln
@@ -67,6 +67,8 @@ EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starsky.feature.metaupdate", "starsky.feature.metaupdate\starsky.feature.metaupdate.csproj", "{FBB448C3-A995-469C-8004-11FA006F65C5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starsky.feature.export", "starsky.feature.export\starsky.feature.export.csproj", "{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -161,6 +163,10 @@ Global
 		{FBB448C3-A995-469C-8004-11FA006F65C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FBB448C3-A995-469C-8004-11FA006F65C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FBB448C3-A995-469C-8004-11FA006F65C5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -190,5 +196,6 @@ Global
 		{81C4FBDB-B778-4A98-AEE4-FC7B6E341589} = {4B9276C3-651E-48D3-B3A7-3F4D74F3D01A}
 		{0F47FD1D-CE79-4069-AFD4-252ED9D9445E} = {C757587F-8490-4815-97CE-9BE39A6C82C6}
 		{FBB448C3-A995-469C-8004-11FA006F65C5} = {4B9276C3-651E-48D3-B3A7-3F4D74F3D01A}
+		{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6} = {4B9276C3-651E-48D3-B3A7-3F4D74F3D01A}
 	EndGlobalSection
 EndGlobal
